requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
diff-match-patch==20241021
Django==5.1.3
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==4.2.1
et_xmlfile==2.0.0
executing==2.1.0
ipython==8.29.0
jedi==0.19.2
matplotlib-inline==0.1.7
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
psycopg==3.2.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
tablib==3.7.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
